+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+# 0.7.0 - 2022-05-xx
+* Add `createCourseTemplate` route
+* Add generic `makeRequest` route
 # 0.6.7 - 2022-01-25
 * Add offset parameter in `importCourse` API